Elastic: Add elasticsearch-curator for index management
The elasticsearch-curator chart is already part of
the helm/stable charts pulled in as part of monitor-helm.
The chart is patched up to the latest version which includes
the k8s 1.16 support.

Three actions are set in the manifest to delete old filebeat,
metricbeat and collectd indices according to sizes set through
environment variables which can be overridden.

Story: 2005733
Task: 37148

Change-Id: Ifad8a859583c898e47c62589381119b5fbc0b50f
Signed-off-by: Kevin Smith <kevin.smith@windriver.com>