elastic: update masterService in manifest
For dual controller configs with no workers, the
elasticsearch-data pods will be configured as master-capable.
This means the corresponding service should be listed in the
discovery.seed_hosts along with the master service.
Use the template's masterService setting to configure this.

For configs with 3 master pods and the data pods are not master-
capable this setting will not hurt.

Without this setting, if the sole master pod went down for any
reason including manual delete or application-update, it would
not be able to rejoin the cluster.

Change-Id: I19302956bf079850459125ea4dd73a0d9c4365cb
Story: 2005733
Task: 37259
Signed-off-by: Kevin Smith <kevin.smith@windriver.com>