elastic: application-update support
Modify the podManagementPolicy for elasticsearch pods
to OrderedReady so no outage is incurred during application-updates.
This setting will also affect initial application-apply, so the elasticsearch
helm chart is also modified to set the initial_master_nodes to
just the first master pod so it will go ready and the apply can proceed.

Change-Id: I87e73f6000430f81f2dc08d37ecd523682377ce3
Story: 2005733
Task: 37290
Signed-off-by: Kevin Smith <kevin.smith@windriver.com>