Adjust curator limits based on num of data nodes
Add an initcontainer to elasticsearch-curator that will
poll elasticsearch for the number of data nodes.  If the number
of data nodes found is 1, indicating either a locked controller
node or AIO-SX, write a file with a LIMIT_FACTOR of 2.  This
file will be sourced and used to adjust the existing index
limits accordingly before curator is run in the main container.

As the curator image does not include curl, use the filebeat
image, which also requires mounting a dummy directory:
/usr/share/filebeat.

Change-Id: I96ac6ee02e1446ca251a4c0cce5397669e399cea
Closes-Bug: 1852472
Partial-Bug: 1852462
Signed-off-by: Kevin Smith <kevin.smith@windriver.com>