Remove unused image references from monitor app
Remove all references to images that are not used by the app
to prevent downloading them. Removals include images for
testing kibana and enabling logstash exporter.

Also explicitly add a reference to the nginx defaultbackend image
in the manifest as it fails to appear in the image download
list in the application images yaml file.  This is possibly an
image parsing issue to be looked at later.

Change-Id: Id6ec19e6c8e5df4aa543a3e4f6b3c36b954d0971
Partial-Bug: 1856197
Signed-off-by: Kevin Smith <kevin.smith@windriver.com>
(cherry picked from commit f6010f4ed211817a0f0fab8ef9aa175085068f0e)