stx-monitor elasticsearch readiness probe enhancements
Modify the readiness probe in the elasticsearch chart so that
2 sets of cluster health parameters can be set.  One set of cluster
health parameters will be used when no connection to the master
is possible or there are less than 2 elasticsearch data nodes in the
cluster.  The other will be used when connection to the master is
possible and 2 or more elasticsearch data nodes are present in the
cluster.

This is necessary as in various cluster recovery scenarios and AIO-SX
only local node health should be checked, whereas when modifying
elasticsearch data nodes we need to ensure no recovery is in progress
before advancing to updating the next data node.

Change-Id: I8125c3c5b87e081a00907c519e3d513c74031c70
Closes-Bug: 1869001
Signed-off-by: Kevin Smith <kevin.smith@windriver.com>