Fix Elasticsearch readiness probe http endpoint
The readiness probe endpoint was changed in the 7.6.0 chart update. This
prevents the application from being reapplied after being removed since
elasticsearch stores its state in the persistent volume, meaning that
when it is reapplied it will try and fail to achieve quorum. Changing
the readiness probe to make a local healthcheck lets it advance to the
next master pod, instead of getting stuck on the first master pod.

Story: 2007221
Task: 39491

Change-Id: I720cec5a7a50a0207669258b97497aba9c99efee
Signed-off-by: Simon Cousineau <Simon.Cousineau@windriver.com>